Address true-on-policy debug helper review nits
